project: fix problems regarding padding, xd

+++ eBPF_code/common_kern_user_datastructure.h
@@ -2,17 +2,19 @@
 #define __COMMON_KERN_USER_H_DATASTRUCTURE
 
 /* This is the data record stored in the map */
-struct flow { // Stored in BIG ENDIAN!
-        __be32  saddr;
-        __be32  daddr;
-        __be16  sport;
-        __be16  dport;
-        __u8    protocol;
+struct flow
+{ // Stored in BIG ENDIAN!
+        __be32 saddr;
+        __be32 daddr;
+        __be16 sport;
+        __be16 dport;
+        __u8 protocol;
 };
 
-struct info_map {
-	__u32 packets;
-	__u64 bytes;
+struct info_map
+{
+        __u64 packets;
+        __u64 bytes;
 };
 
 #ifndef XDP_ACTION_MAX
